feat: 구현 NewTryCountElement
Ref: #17
